migrations/models/1_20221115004650_add_field_desc.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `role` ADD `desc` LONGTEXT NOT NULL  COMMENT 'text';"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `role` DROP COLUMN `desc`;"""
