requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.2
black==22.10.0
click==8.1.3
dictdiffer==0.9.0
fastapi==0.87.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pydantic==1.10.2
PyMySQL==1.0.2
pypika-tortoise==0.1.6
pytz==2022.6
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.19.0
